+++ example/main.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:table_sticky_headers/table_sticky_headers.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(TableSimple());
 
-class MyApp extends StatelessWidget {
+class TableSimple extends StatelessWidget {
+  final List<String> columns = List.generate(10, (i) => '${i + 1}');
+  final List<String> rows = List.generate(20, (i) => '${i + 1}');
+
   @override
   Widget build(BuildContext context) {
-    List<String> tops = List.generate(20, (i) => '${i + 1}');
-    List<String> lefts = List.generate(20, (i) => '${i + 1}');
-
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
@@ -16,14 +16,15 @@ class MyApp extends StatelessWidget {
           backgroundColor: Colors.amber,
         ),
         body: StickyHeadersTable(
-          columnsLength: tops.length,
-          rowsLength: lefts.length,
-          columnsTitleBuilder: (i) => 'Top ${tops[i]}',
-          rowsTitleBuilder: (i) => 'Left ${lefts[i]}',
-          contentCellBuilder: (i, j) => 'T${tops[i]} : L${lefts[j]}',
-          legendCell: 'Permanently Sticky',
+          columnsLength: columns.length,
+          rowsLength: rows.length,
+          columnsTitleBuilder: (i) => Text('Top ${columns[i]}'),
+          rowsTitleBuilder: (i) => Text('Left ${rows[i]}'),
+          contentCellBuilder: (i, j) => Text('T${columns[i]} : L${rows[j]}'),
+          legendCell: Text('Sticky Legend'),
         ),
       ),
     );
   }
 }
+