+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Release Date
 * 27.10.2019
 
+## [1.1.0] -
+* Changed table builders input parameters so they take Widget instead of text. Now user has full control of setting table content.
+* Added example to show how to use table decoration, including click listeners.
+
 ## [1.0.1] -
 * Added demo image